+++ packages/jest-cli/src/cli/args.js
@@ -20,12 +20,22 @@ export const check = (argv: Argv) => {
     );
   }
 
-  if (argv.onlyChanged && argv.watchAll) {
-    throw new Error(
-      'Both --onlyChanged and --watchAll were specified, but these two ' +
-        'options do not make sense together. Try the --watch option which ' +
-        'reruns only tests related to changed files.',
-    );
+  for (const key of [
+    'onlyChanged',
+    'lastCommit',
+    'changedFilesWithAncestor',
+    'changedFilesToContributeTo',
+  ]) {
+    if (argv[key]) {
+      argv.onlyChanged = true;
+    }
+    if (argv[key] && argv.watchAll) {
+      throw new Error(
+        `Both --${key} and --watchAll were specified, but these two ` +
+          'options do not make sense together. Try the --watch option which ' +
+          'reruns only tests related to changed files.',
+      );
+    }
   }
 
   if (argv.findRelatedTests && argv._.length === 0) {
@@ -102,10 +112,18 @@ export const options = {
       ' dependency information.',
     type: 'string',
   },
+  changedFilesToContributeTo: {
+    description:
+      'Runs tests related the changes since the provided branch. If the ' +
+      'current branch has diverged from the given branch, then only changes ' +
+      'made locally will be tested. Behaves similarly to `--onlyChanged`.',
+    nargs: 1,
+    type: 'string',
+  },
   changedFilesWithAncestor: {
     description:
-      'When used together with `--onlyChanged` or `--watch`, it runs tests ' +
-      'related to the current changes and the changes made in the last commit. ',
+      'Runs tests related to the current changes and the changes made in the ' +
+      'last commit. Behaves similarly to `--onlyChanged`.',
     type: 'boolean',
   },
   ci: {
@@ -267,8 +285,8 @@ export const options = {
   lastCommit: {
     default: undefined,
     description:
-      'When used together with `--onlyChanged`, it will run all tests ' +
-      'affected by file changes in the last commit made.',
+      'Run all tests affected by file changes in the last commit made. ' +
+      'Behaves similarly to `--onlyChanged`.',
     type: 'boolean',
   },
   listTests: {
@@ -353,7 +371,7 @@ export const options = {
     description:
       'Attempts to identify which tests to run based on which ' +
       "files have changed in the current repository. Only works if you're " +
-      'running tests in a git repository at the moment.',
+      'running tests in a git or hg repository at the moment.',
     type: 'boolean',
   },
   onlyFailures: {