+++ types/Config.js
@@ -76,6 +76,7 @@ export type InitialOptions = {
   cacheDirectory?: Path,
   clearMocks?: boolean,
   changedFilesWithAncestor?: boolean,
+  changedFilesToContributeTo?: string,
   collectCoverage?: boolean,
   collectCoverageFrom?: Array<Glob>,
   collectCoverageOnlyFrom?: {[key: string]: boolean},
@@ -157,6 +158,7 @@ export type SnapshotUpdateState = 'all' | 'new' | 'none';
 
 export type GlobalConfig = {|
   bail: boolean,
+  changedFilesToContributeTo: string,
   changedFilesWithAncestor: boolean,
   collectCoverage: boolean,
   collectCoverageFrom: Array<Glob>,