+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * `[jest-cli]` Make Jest exit without an error when no tests are found in the
   case of `--lastCommit`, `--findRelatedTests`, or `--onlyChanged` options
   having been passed to the CLI
+* `[jest-cli]` Allow selectively running tests for code changed since arbitrary
+  revisions. ([#5188](https://github.com/facebook/jest/pull/5188))
 
 ### Fixes
 * `[jest-cli]` Use `import-local` to support global Jest installations.
@@ -13,6 +15,8 @@
   ([#5289](https://github.com/facebook/jest/pull/5289))
 * `[docs]` Update mention of the minimal version of node supported [#4947](https://github.com/facebook/jest/issues/4947)
 * `[jest-cli]` Fix missing newline in console message ([#5308](https://github.com/facebook/jest/pull/5308))
+* `[jest-cli]` `--lastCommit` and `--changedFilesWithAncestor` now take effect
+  even when `--onlyChanged` is not specified. ([#5307](https://github.com/facebook/jest/pull/5307))
 
 ### Chore & Maintenance